Uploaded Backend & Frontend
The Backend is in FastAPI & the Frontend is In React. Made the project Scalable with Proper Comments and Directory Structure. The Background is Firebase. needed Additional 12 hours in order to finish. The Project Consists of 3 users  Admin , Police , Client where this Commit is for Admin and Police another Commit will be done For Android App. Linking is Left.